YoutubeConverter/requirements.txt
customtkinter
darkdetect
Pillow
yt-dlp
requests
humanize
selenium
webdriver-manager
